slack/pyproject.toml
[project]
name = "astro-bot-v1"
version = "0.1.0"
description = "Slack manifest"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]
